separate side tools from runtime lib
